// types/gesellschaft.ts
export interface Gesellschaft {
    // organizer_desc?: string;
    // organizer_email?: string;
    // organizer_link?: string;
    // organizer_name?: string;
    // organizer_tax?: number;

    permalink?: string;
    desc?: string;
    email?: string;
    link?: string;
    name: string;
    tax: number;
}